+++ src/deser/stcs.rs
@@ -21,8 +21,8 @@ use stc_s::{
 };
 
 use healpix::nested::{
-  bmoc::{BMOCBuilderUnsafe, BMOC},
-  box_coverage, cone_coverage_approx_custom, elliptical_cone_coverage_custom, polygon_coverage,
+  bmoc::BMOC, box_coverage, cone_coverage_approx_custom, elliptical_cone_coverage_custom,
+  polygon_coverage, zone_coverage,
 };
 
 use crate::{moc::range::RangeMOC, qty::Hpx};
@@ -83,8 +83,7 @@ impl CompoundVisitor for Stc2Moc {
   type Error = Stc2MocError;
 
   fn visit_allsky(&mut self) -> Result<Self::Value, Self::Error> {
-    // Ok(BMOC::new_allsky(self.depth))
-    Ok(new_allsky(self.depth))
+    Ok(BMOC::new_allsky(self.depth))
   }
 
   fn visit_circle(&mut self, circle: &CircleParams) -> Result<Self::Value, Self::Error> {
@@ -320,8 +319,38 @@ impl SpaceVisitor for Stc2Moc {
     Err(Stc2MocError::SimplePositionNotSupported)
   }
 
-  fn visit_position_interval(self, _: &PositionInterval) -> Result<Self::Value, Self::Error> {
-    Err(Stc2MocError::PositionIntervalNotSupported)
+  fn visit_position_interval(
+    self,
+    interval: &PositionInterval,
+  ) -> Result<Self::Value, Self::Error> {
+    // We use compound visitor onyl to check interval parameters
+    self.new_compound_visitor(&interval.pre, &interval.post)?;
+    let depth = self.depth;
+    let corners = interval
+      .lo_hi_limits
+      .iter()
+      .step_by(2)
+      .zip(interval.lo_hi_limits.iter().skip(1).step_by(2))
+      .map(|(lon_deg, lat_deg)| {
+        let lon = lon_deg2rad(*lon_deg)?;
+        let lat = lat_deg2rad(*lat_deg)?;
+        Ok((lon, lat))
+      })
+      .collect::<Result<Vec<(f64, f64)>, Stc2MocError>>()?;
+    let mut corners_it = corners
+      .iter()
+      .cloned()
+      .step_by(2)
+      .zip(corners.iter().cloned().skip(1).step_by(2));
+    if let Some(((ra_min, dec_min), (ra_max, dec_max))) = corners_it.next() {
+      let mut bmoc = zone_coverage(depth, ra_min, dec_min, ra_max, dec_max);
+      while let Some(((ra_min, dec_min), (ra_max, dec_max))) = corners_it.next() {
+        bmoc = bmoc.or(&zone_coverage(depth, ra_min, dec_min, ra_max, dec_max));
+      }
+      Ok(bmoc.into())
+    } else {
+      Ok(RangeMOC::<u64, Hpx<u64>>::new_empty(depth))
+    }
   }
 
   fn visit_allsky(self, bmoc: BMOC) -> Result<Self::Value, Self::Error> {
@@ -386,13 +415,6 @@ fn lat_deg2rad(lat_deg: f64) -> Result<f64, Stc2MocError> {
   }
 }
 
-// TODO: remove when newt version of cdshealpix will be published
-fn new_allsky(depth: u8) -> BMOC {
-  let mut builder = BMOCBuilderUnsafe::new(depth, 12);
-  builder.push_all(0, 0, 11, true);
-  builder.to_bmoc()
-}
-
 /// Create new S-MOC from the given STC-S string.
 ///
 /// # WARNING
@@ -453,7 +475,7 @@ mod tests {
   use super::stcs2moc;
 
   #[test]
-  fn test_from_stcs() {
+  fn test_from_stcs_circle() {
     let moc1 = stcs2moc(10, Some(2), "Circle ICRS 147.6 69.9 0.4").unwrap();
     let moc2 = RangeMOC::<u64, Hpx<u64>>::from_cone(
       147.6_f64.to_radians(),
@@ -463,5 +485,121 @@ mod tests {
       2,
     );
     assert_eq!(moc1, moc2)
+  }
+
+  #[test]
+  fn test_from_stcs_allsky() {
+    let moc1 = stcs2moc(10, Some(2), "Allsky ICRS").unwrap();
+    let moc2 = RangeMOC::<u64, Hpx<u64>>::new_full_domain(10);
+    assert_eq!(moc1, moc2);
+    let moc1 = stcs2moc(18, Some(2), "Allsky ICRS").unwrap();
+    let moc2 = RangeMOC::<u64, Hpx<u64>>::new_full_domain(18);
+    assert_eq!(moc1, moc2);
+  }
+
+  #[test]
+  fn test_from_stcs_not_allsky() {
+    let moc1 = stcs2moc(15, Some(2), "Not ICRS(Allsky)").unwrap();
+    let moc2 = RangeMOC::<u64, Hpx<u64>>::new_empty(15);
+    assert_eq!(moc1, moc2);
+  }
+
+  #[test]
+  fn test_from_stcs_union_difference_intersection() {
+    let stcs1 = r#"
+Difference ICRS (
+    Polygon 272.536719 -19.461249 272.542612 -19.476380 272.537389 -19.491509 272.540192 -19.499823
+            272.535455 -19.505218 272.528024 -19.505216 272.523437 -19.500298 272.514082 -19.503376
+            272.502271 -19.500966 272.488647 -19.490390  272.481932 -19.490913 272.476737 -19.486589
+            272.487633 -19.455645 272.500386 -19.444996 272.503003 -19.437557 272.512303 -19.432436
+            272.514132 -19.423973 272.522103 -19.421523 272.524511 -19.413250 272.541021 -19.400024
+            272.566264 -19.397500 272.564202 -19.389111 272.569055 -19.383210 272.588186 -19.386539
+            272.593376 -19.381832 272.596327 -19.370541 272.624911 -19.358915 272.629256 -19.347842
+            272.642277 -19.341020 272.651322 -19.330424 272.653174 -19.325079 272.648903 -19.313708
+            272.639616 -19.311098 272.638128 -19.303083 272.632705 -19.299839 272.627971 -19.289408
+            272.628226 -19.276293 272.633750 -19.270590 272.615109 -19.241810 272.614704 -19.221196
+            272.618224 -19.215572 272.630809 -19.209945 272.633540 -19.198681 272.640711 -19.195292
+            272.643028 -19.186751 272.651477 -19.182729 272.649821 -19.174859 272.656782 -19.169272
+            272.658933 -19.161883 272.678012 -19.159481 272.689173 -19.176982 272.689395 -19.183512
+            272.678006 -19.204016 272.671112 -19.206598 272.664854 -19.203523 272.662760 -19.211156
+            272.654435 -19.214434 272.652969 -19.222085 272.656724 -19.242136 272.650071 -19.265092
+            272.652868 -19.274296 272.660871 -19.249462 272.670041 -19.247807 272.675533 -19.254935
+            272.673291 -19.273917 272.668710 -19.279245 272.671460 -19.287043 272.667507 -19.293933
+            272.669261 -19.300601 272.663969 -19.307130 272.672626 -19.308954 272.675225 -19.316490
+            272.657188 -19.349105 272.657638 -19.367455 272.662447 -19.372035 272.662232 -19.378566
+            272.652479 -19.386871 272.645819 -19.387933 272.642279 -19.398277 272.629282 -19.402739
+            272.621487 -19.398197 272.611782 -19.405716 272.603367 -19.404667 272.586162 -19.422703
+            272.561792 -19.420008 272.555815 -19.413012 272.546500 -19.415611 272.537427 -19.424213
+            272.533081 -19.441402
+    Union (
+        Polygon 272.511081 -19.487278 272.515300 -19.486595 272.517029 -19.471442
+                272.511714 -19.458837 272.506430 -19.459001 272.496401 -19.474322 272.504821 -19.484924
+        Polygon 272.630446 -19.234210 272.637274 -19.248542 272.638942 -19.231476 272.630868 -19.226364
+    )
+)"#;
+    let stcs2 = r#"
+Intersection ICRS (
+    Polygon 272.536719 -19.461249 272.542612 -19.476380 272.537389 -19.491509 272.540192 -19.499823
+            272.535455 -19.505218 272.528024 -19.505216 272.523437 -19.500298 272.514082 -19.503376 
+            272.502271 -19.500966 272.488647 -19.490390  272.481932 -19.490913 272.476737 -19.486589 
+            272.487633 -19.455645 272.500386 -19.444996 272.503003 -19.437557 272.512303 -19.432436 
+            272.514132 -19.423973 272.522103 -19.421523 272.524511 -19.413250 272.541021 -19.400024 
+            272.566264 -19.397500 272.564202 -19.389111 272.569055 -19.383210 272.588186 -19.386539 
+            272.593376 -19.381832 272.596327 -19.370541 272.624911 -19.358915 272.629256 -19.347842 
+            272.642277 -19.341020 272.651322 -19.330424 272.653174 -19.325079 272.648903 -19.313708 
+            272.639616 -19.311098 272.638128 -19.303083 272.632705 -19.299839 272.627971 -19.289408 
+            272.628226 -19.276293 272.633750 -19.270590 272.615109 -19.241810 272.614704 -19.221196 
+            272.618224 -19.215572 272.630809 -19.209945 272.633540 -19.198681 272.640711 -19.195292 
+            272.643028 -19.186751 272.651477 -19.182729 272.649821 -19.174859 272.656782 -19.169272 
+            272.658933 -19.161883 272.678012 -19.159481 272.689173 -19.176982 272.689395 -19.183512 
+            272.678006 -19.204016 272.671112 -19.206598 272.664854 -19.203523 272.662760 -19.211156 
+            272.654435 -19.214434 272.652969 -19.222085 272.656724 -19.242136 272.650071 -19.265092
+            272.652868 -19.274296 272.660871 -19.249462 272.670041 -19.247807 272.675533 -19.254935 
+            272.673291 -19.273917 272.668710 -19.279245 272.671460 -19.287043 272.667507 -19.293933
+            272.669261 -19.300601 272.663969 -19.307130 272.672626 -19.308954 272.675225 -19.316490
+            272.657188 -19.349105 272.657638 -19.367455 272.662447 -19.372035 272.662232 -19.378566
+            272.652479 -19.386871 272.645819 -19.387933 272.642279 -19.398277 272.629282 -19.402739
+            272.621487 -19.398197 272.611782 -19.405716 272.603367 -19.404667 272.586162 -19.422703
+            272.561792 -19.420008 272.555815 -19.413012 272.546500 -19.415611 272.537427 -19.424213
+            272.533081 -19.441402 
+    Not (Polygon 272.511081 -19.487278 272.515300 -19.486595 272.517029 -19.471442 
+                 272.511714 -19.458837 272.506430 -19.459001 272.496401 -19.474322 272.504821 -19.484924)
+    Not (Polygon 272.630446 -19.234210 272.637274 -19.248542 272.638942 -19.231476 272.630868 -19.226364)
+)"#;
+    let moc1 = stcs2moc(16, Some(2), stcs1).unwrap();
+    let moc2 = stcs2moc(16, Some(2), stcs2).unwrap();
+    assert_eq!(moc1, moc2);
+
+    // Write file to check it manually
+    /*
+    use crate::moc::{RangeMOCIntoIterator, RangeMOCIterator};
+    use std::{fs::File, io::BufWriter, path::PathBuf};
+    let path_buf1 = PathBuf::from("resources/stcs/eso.res.moc.fits");
+    let file = File::create(&path_buf1).unwrap();
+    let writer = BufWriter::new(file);
+    moc1
+      .into_range_moc_iter()
+      .to_fits_ivoa(None, None, writer)
+      .unwrap();
+     */
+  }
+
+  #[test]
+  fn test_from_stcs_position_interval() {
+    let moc = stcs2moc(
+      10,
+      Some(2),
+      "PositionInterval ICRS 170 -20 190 10 Resolution 0.0001",
+    )
+    .unwrap();
+    assert_eq!(moc.len(), 1015);
+
+    let moc = stcs2moc(
+      10,
+      Some(2),
+      "PositionInterval ICRS 170 -20 190 10 200 10 210 20 Resolution 0.0001",
+    )
+    .unwrap();
+    assert_eq!(moc.len(), 1383);
   }
 }